APIS-5608 - Move stride module to apiplatform modules
