+++ app/uk/gov/hmrc/apigatekeeperapprovalsfrontend/services/ApplicationActionService.scala
@@ -25,7 +25,7 @@ import uk.gov.hmrc.http.HeaderCarrier
 
 import uk.gov.hmrc.apigatekeeperapprovalsfrontend.controllers.models.ApplicationRequest
 import uk.gov.hmrc.apigatekeeperapprovalsfrontend.domain.models.ApplicationId
-import uk.gov.hmrc.modules.stride.controllers.models.LoggedInRequest
+import uk.gov.hmrc.apiplatform.modules.stride.controllers.models.LoggedInRequest
 
 @Singleton
 class ApplicationActionService @Inject()(
