+++ app/uk/gov/hmrc/apiplatform/modules/stride/config/StrideAuthConfigProvider.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.modules.stride.config
+package uk.gov.hmrc.apiplatform.modules.stride.config
 
 import javax.inject.{Inject, Provider, Singleton}
