+++ conf/application.conf
@@ -40,7 +40,7 @@ play.modules.enabled += "uk.gov.hmrc.mongo.play.PlayMongoModule"
 
 # Default http client
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.HttpClientModule"
-play.modules.enabled += "uk.gov.hmrc.modules.stride.config.ConfigurationModule"
+play.modules.enabled += "uk.gov.hmrc.apiplatform.modules.stride.config.ConfigurationModule"
 play.modules.enabled += "uk.gov.hmrc.apigatekeeperapprovalsfrontend.config.ConfigurationModule"
 
 # Custom error handler
