+++ test/uk/gov/hmrc/apigatekeeperapprovalsfrontend/controllers/ApplicationControllerSpec.scala
@@ -24,10 +24,10 @@ import play.api.inject.guice.GuiceApplicationBuilder
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import uk.gov.hmrc.apigatekeeperapprovalsfrontend.config.AppConfig
-import uk.gov.hmrc.modules.stride.config.StrideAuthConfig
+import uk.gov.hmrc.apiplatform.modules.stride.config.StrideAuthConfig
 import play.api.mvc.MessagesControllerComponents
-import uk.gov.hmrc.modules.stride.connectors.mocks.ApplicationActionServiceMockModule
-import uk.gov.hmrc.modules.stride.connectors.mocks.AuthConnectorMockModule
+import uk.gov.hmrc.apiplatform.modules.stride.connectors.mocks.ApplicationActionServiceMockModule
+import uk.gov.hmrc.apiplatform.modules.stride.connectors.mocks.AuthConnectorMockModule
 import uk.gov.hmrc.apiplatform.modules.submissions.services.SubmissionServiceMockModule
 import uk.gov.hmrc.apigatekeeperapprovalsfrontend.config.ErrorHandler
 import uk.gov.hmrc.apigatekeeperapprovalsfrontend.domain.models.{ApplicationId,Application}
